catchup arithmetic modules
