--- quil-py/quil/__init__.pyi
@@ -1,3 +1,0 @@
-import instructions as instructions
-import program as program
-import validation as validation
