--- quil-py/quil/instructions/_memory_region.pyi
@@ -1,9 +1,0 @@
-class MemoryReference:
-    @property
-    def name(self) -> str: ...
-    @name.setter
-    def name(self, name: str): ...
-    @property
-    def index(self) -> int: ...
-    @index.setter
-    def index(self, index: int): ...
